Logo
Logo is in and aligned. I have left it as is because I couldn't exactly figure out how to get the text to center itself with the image. Feel free to see if you can make it better to look at, but I feel as if I have done all I could at this point.